requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
Werkzeug==3.0.1 